Extract model related codes to separate model components
- Doing this because we want to be able to call models at their route, otherwise, the FormControl does not take focus and the form becomes unresponsive.
- Todo: This would open multiple models if the same model component is in the same tree of route. Happens in case of Insurance policy. Have to fix it by tagging models so that only models that are supposed to open would actually open.
