+++ src/widgets/FormInputs/index.js
@@ -1,0 +1,5 @@
+export { FormTextInput } from './FormTextInput';
+export { FormDateInput } from './FormDateInput';
+export { FormToggle } from './FormToggle';
+export { FormDropdown } from './FormDropdown';
+export { FormSlider } from './FormSlider';
