+++ src/widgets/modalChildren/index.js
@@ -9,3 +9,4 @@ export { ConfirmForm } from './ConfirmForm';
 export { GenericChoiceList } from './GenericChoiceList';
 export { MultiSelectList } from './MultiSelectList';
 export { ToggleSelector } from './ToggleSelector';
+export { PatientEditModal } from './PatientEditModal';
